third_party/WebKit/LayoutTests/fast/dom/shadow/shadowdom-for-unsupported-elements.html

<!DOCTYPE html>
<body>
<script src="../../js/resources/js-test-pre.js"></script>
<script>
description('We do not allow adding author shadow trees to some types of elements.');
// See crbug.com/234020
var elements = [
    'audio', 'canvas', 'details', 'embed', 'fieldset', 'iframe', 'img', 'input',
    'keygen', 'meter', 'object', 'progress', 'select', 'summary', 'textarea', 'video'
];
var element;
elements.forEach(function(elementName) {
    element = document.createElement(elementName);
    debug(elementName);
    shouldThrow('element.createShadowRoot()');
});

debug('svg');
element = document.createElementNS('http://www.w3.org/2000/svg', 'svg');
shouldThrow('element.createShadowRoot()');

</script>
<script src="../../js/resources/js-test-post.js"></script>
</body>
